+++ src/api-adapter/react/create-api.js
@@ -1,3 +1,5 @@
+var stringifyWithLimit = require('./stringify-with-limit.js');
+
 var hasOwnProperty = Object.prototype.hasOwnProperty;
 var elementMap = {};
 var getNode;
@@ -56,7 +58,7 @@ function getComponentNameByNode(node) {
     }
   }
 
-  return name ? '<' + name + '>' : 'Unknown';
+  return name || 'Unknown';
 }
 
 
@@ -119,12 +121,58 @@ function isComponentRootNode(node) {
   );
 }
 
+function getPropValueType(value) {
+  if (typeof value === 'string') {
+    return 'string';
+  } else if (typeof value === 'function') {
+    return 'function';
+  } else {
+    return 'other';
+  }
+}
+
+function getPropValueText(type, value) {
+  switch (type) {
+    case 'string':
+      return stringifyWithLimit.sliceText(value, 20);
+
+    case 'function':
+      // Additional 9 characters for 'function '
+      return stringifyWithLimit(value, 29);
+
+    case 'other':
+      return stringifyWithLimit(value, 20);
+  }
+}
+
+function getInstancePropsInfo(instanceProps, getLocation) {
+  var props = [];
+
+  for (var prop in instanceProps) {
+    if (hasOwnProperty.call(instanceProps, prop)) {
+      var value = instanceProps[prop];
+      var type = getPropValueType(value);
+      var valueText = getPropValueText(type, value);
+
+      props.push({
+        key: prop,
+        type: type,
+        valueText: valueText,
+        valueLoc: value && (typeof value === 'object' || typeof value === 'function') ? getLocation(value) : undefined
+      });
+    }
+  }
+
+  return props;
+}
+
 function getAdditionalInstanceInfo(element) {
   var instanceRootNode = this.getInstanceRootNode(element);
   var instance = element && element._instance;
   var cls = instance && instance.constructor;
   var wrapperClassMap = {};
   var decorators = [];
+  var props = [];
   var info;
 
   if (typeof cls == 'function') {
@@ -147,7 +195,7 @@ function getAdditionalInstanceInfo(element) {
 
       if (wrapperClass) {
         classLoc = this.getLocation(wrapperClass);
-        renderLoc = this.getLocation(wrapperClass && wrapperClass.prototype.render);
+        renderLoc = this.getClassMethodLocation(wrapperClass, 'render');
       }
 
       return {
@@ -160,13 +208,18 @@ function getAdditionalInstanceInfo(element) {
     }, this);
   }
 
+  if (instance.props) {
+    props = getInstancePropsInfo(instance.props, this.getLocation);
+  }
+
   return [
     {
       name: 'Instance',
       childNodes: [
         createInstanceView({
           name: this.getComponentNameByNode(instanceRootNode),
-          loc: this.getNestedComponentNodeLocation(instanceRootNode)
+          loc: this.getNestedComponentNodeLocation(instanceRootNode),
+          childNodes: props
         })
       ]
     },
@@ -176,6 +229,7 @@ function getAdditionalInstanceInfo(element) {
         createClassView({
           cls: cls,
           isClass: info && info.type === 'class',
+          getClassMethodLocation: this.getClassMethodLocation,
           getLocation: this.getLocation,
           childNodes: decorators
         })
@@ -190,21 +244,16 @@ function getInstanceLocation(element) {
 
 function getInstanceRenderLocation(element) {
   var instance = element && element._instance;
-  var render;
 
   if (instance) {
     if (hasOwnProperty.call(instance, 'render')) {
-      render = instance.render;
+      return this.getLocation(instance.render);
     } else {
       var constructor = unwrap(instance.constructor, this.getDevInfo);
-      if (constructor && constructor.prototype) {
-        render = constructor.prototype.render;
+      if (constructor) {
+        return this.getClassMethodLocation(constructor, 'render');
       }
     }
-  }
-
-  if (typeof render == 'function') {
-    return this.getLocation(render);
   }
 }